% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gps.R
\name{estimate_gps}
\alias{estimate_gps}
\title{Title}
\usage{
estimate_gps(
  formula,
  data,
  method = "multinom",
  reference = NULL,
  by = NULL,
  missing = NULL,
  fit.object = FALSE,
  verbose.output = FALSE,
  ...
)
}
\description{
Title
}
