% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_gps.R
\name{make_opt_args}
\alias{make_opt_args}
\title{Defining custom s3 class to create and validate opt_args input}
\usage{
make_opt_args(
  data = NULL,
  formula,
  reference = NULL,
  gps_method = paste0("m", 1:10),
  matching_method = c("fullopt", "nnm"),
  caliper = seq(0.01, 10, 0.01),
  order = c("desc", "asc", "original", "random"),
  cluster = 2,
  replace = c(TRUE, FALSE),
  ties = c(TRUE, FALSE),
  ratio = 1,
  min_controls = 1,
  max_controls = 1,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Defining custom s3 class to create and validate opt_args input
}
