% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic}
\alias{mosaic}
\title{Plot the distribution of categorical covariates}
\usage{
mosaic(
  data = NULL,
  y = NULL,
  group = NULL,
  facet = NULL,
  ncol = 1,
  significance = FALSE,
  plot.name = NULL,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Plot the distribution of categorical covariates
}
