% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{raincloud}
\alias{raincloud}
\title{Title}
\usage{
raincloud(
  data = NULL,
  y = NULL,
  group = NULL,
  facet = NULL,
  limits = NULL,
  jitter = NULL,
  alpha = NULL,
  legend.position = NULL,
  kernel.method = NULL,
  save = FALSE
)
}
\description{
Title
}
