% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmatch-package.R
\docType{package}
\name{vecmatch-package}
\alias{vecmatch}
\alias{vecmatch-package}
\title{vecmatch: Vector Matching for Generalized Propensity Scores}
\description{
The \code{vecmatch} package implements vector matching methods based
on generalized propensity scores (GPS) for balancing multiple treatment
groups, along with diagnostic tools and visualization functions.
}
\section{Life cycle}{

\code{vecmatch} is currently in a \emph{stable} stage. The core matching and
diagnostic functionality is implemented and covered by tests, and the
package is actively maintained. Minor API changes may still occur in
response to user feedback and further methodological development, but
major breaking changes are not anticipated. Future work will focus on
extending the set of matching algorithms, adding further diagnostics,
and improving performance and documentation.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Polymerase3/vecmatch}
  \item \url{https://polymerase3.github.io/vecmatch/}
  \item Report bugs at \url{https://github.com/Polymerase3/vecmatch/issues}
}

}
\author{
\strong{Maintainer}: Mateusz Kolek \email{mati.kolek13@gmail.com} (\href{https://orcid.org/0000-0001-6470-4830}{ORCID}) [copyright holder]

}
\keyword{internal}
