% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{vspine}
\alias{vspine}
\title{Fixing bug in productplots::prodcalc}
\usage{
vspine(...)
}
\value{
No return value, called for side effects
}
\description{
Fixing bug in productplots::prodcalc
}
\keyword{internal}
