% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gps.R
\name{estimate_gps}
\alias{estimate_gps}
\title{Estimate Generalized Propensity Scores with various methods}
\usage{
estimate_gps(
  formula,
  data = NULL,
  method = "multinom",
  reference = NULL,
  by = NULL,
  missing = NULL,
  fit.object = FALSE,
  verbose.output = FALSE,
  ...
)
}
\arguments{
\item{formula}{description}

\item{data}{description}

\item{method}{description}

\item{reference}{description}

\item{by}{description}

\item{missing}{description}

\item{fit.object}{description}

\item{verbose.output}{description}

\item{...}{description}
}
\description{
Estimate Generalized Propensity Scores with various methods
}
\details{
Additional details...
}
