% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_gps.R
\name{get_select_params}
\alias{get_select_params}
\title{Extract Parameter Grid for Selected Configurations}
\usage{
get_select_params(x, smd_group = NULL)
}
\arguments{
\item{x}{A \code{select_result} object returned by \code{select_opt()}.}

\item{smd_group}{Optional character vector of SMD groups
(e.g. \code{"0.10-0.15"}). If provided, the returned data.frame
is subsetted to these groups. If \code{NULL}, all rows are returned.}
}
\value{
A data.frame with the parameter combinations corresponding
to the selected configurations.
}
\description{
Extract Parameter Grid for Selected Configurations
}
