% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_gps.R
\name{match_gps}
\alias{match_gps}
\title{Title}
\usage{
match_gps(
  treatment = NULL,
  data = NULL,
  caliper = NULL,
  reference = NULL,
  by = NULL,
  combos = NULL,
  kmeans.args = NULL,
  ratio = 1,
  replace = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
