% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{raincloud}
\alias{raincloud}
\title{Title}
\usage{
raincloud(
  data = NULL,
  y = NULL,
  group = NULL,
  facet = NULL,
  ncol = 1,
  significance = FALSE,
  limits = NULL,
  jitter = 0.1,
  alpha = 0.4,
  save = FALSE,
  plot.name = NULL,
  overwrite = FALSE
)
}
\description{
Title
}
