% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raincloud.R
\name{raincloud}
\alias{raincloud}
\title{Examine the Distribution of Continuous Data with Raincloud Plots}
\usage{
raincloud(
  data = NULL,
  y = NULL,
  group = NULL,
  facet = NULL,
  ncol = 1,
  significance = FALSE,
  limits = NULL,
  jitter = 0.1,
  alpha = 0.4,
  save = FALSE,
  plot.name = NULL,
  overwrite = FALSE
)
}
\description{
The \code{raincloud()} function allows to generate raincloud plots for continuous
data in an easy and uncomplicated way. Raincloud plots consist of three main
elements:
/item{Distribution plots}{For example density plots, histograms or violin
plots wit the mean values of respective groups}
/item{Jittered point plots}{Depicting the underlying distribution of the data
in the rawest form}
/item{Boxplots}
The function is based on the \code{ggplot2} package, which must already be
preinstalled
}
