% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmatch-package.R
\docType{package}
\name{vecmatch-package}
\alias{vecmatch-package}
\title{vecmatch: Generalized Propensity Score Estimation and Matching for Multiple Groups}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements the Vector Matching algorithm to match multiple treatment groups based on previously estimated generalized propensity scores. The package includes tools for visualizing initial confounder imbalances, estimating treatment assignment probabilities using various methods, defining the common support region, performing matching across multiple groups, and evaluating matching quality. For more details, see Lopez and Gutman (2017) \doi{10.1214/17-STS612}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Polymerase3/vecmatch}
  \item Report bugs at \url{https://github.com/Polymerase3/vecmatch/issues}
}

}
\author{
\strong{Maintainer}: Mateusz Kolek \email{mati.kolek13@gmail.com} (\href{https://orcid.org/0000-0001-6470-4830}{ORCID}) [copyright holder]

}
\keyword{internal}
