% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmatch-package.R
\docType{package}
\name{vecmatch-package}
\alias{vecmatch-package}
\title{vecmatch: Calculate Generalized Propensity Scores and Match Multiple Groups}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of Vector Matching algorithm. Enables the assessment of initial confounder imbalances through elegant graphics, estimates treatment assignment probabilities using various methods, defines the common support region, performs matching across multiple groups and assesses its quality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Polymerase3/vecmatch}
  \item Report bugs at \url{https://github.com/Polymerase3/vecmatch/issues}
}

}
\author{
\strong{Maintainer}: Mateusz Kolek \email{mati.kolek13@gmail.com} (\href{https://orcid.org/0000-0001-6470-4830}{ORCID}) [copyright holder]

}
\keyword{internal}
