% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecmatch-package.R
\docType{package}
\name{vecmatch-package}
\alias{vecmatch-package}
\title{vecmatch: Calculate Genaralized Propensity Scores and Match Multiple Groups}
\description{
An implementation of Vector Matching algorithm. Enables the assesment of initial covariate imbalances through ellegant graphics, estimates treatment assignment probabilities using various methods, defines the common support region, and performs matching across multiple groups.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Polymerase3/vecmatch}
  \item Report bugs at \url{https://github.com/Polymerase3/vecmatch/issues}
}

}
\author{
\strong{Maintainer}: Mateusz Kolek \email{mati.kolek@gmail.com} (\href{https://orcid.org/0000-0001-6470-4830}{ORCID})

}
\keyword{internal}
